pyproject.toml
[tool.poetry]
name = "homework5"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
django-phonenumber-field = "^8.0.0"
pillow = "^11.1.0"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
